safari-600.7-branch/LayoutTests/platform/ios-sim/fast/css/line-height-zoom-get-computed-style-expected.txt

This tests that page zoom and getComputedStyle of line-height works well together. To test manually, open this page in a browser window, and zoom-in and reload repeatedly. This text should get bigger at each step proportionately but the line-height value should stay constant.On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
FAIL 11.67 should be 12. Was 11.67.
FAIL 11.11 should be 12. Was 11.11.
FAIL 11.57 should be 12. Was 11.57.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
FAIL 12.06 should be 12. Was 12.06.
text of font-size 10px with 120% line height.
